Profile Section

+++ database/migrations/2023_03_29_210829_create_services_table.php
@@ -15,10 +15,11 @@ return new class extends Migration
             $table->id();
             $table->string('name');
             $table->string('description');
-            $table->integer('price');
+            $table->decimal('price', 8, 2);
             $table->string('image');
             $table->timestamps();
         });
+
     }
 
     /**
